+++ packages/UsageAnalysis/src/package-api.ts
@@ -220,10 +220,18 @@ export namespace queries {
     return await grok.data.query('UsageAnalysis:Groups', {});
   }
 
+  /**
+   * @param {string} id
+   *   semType: user_id
+   */
   export async function userById(id: string ): Promise<DG.DataFrame> {
     return await grok.data.query('UsageAnalysis:UserById', { id });
   }
 
+  /**
+   * @param {string} email
+   *   semType: email
+   */
   export async function userInfoByEmailPanel(email: string ): Promise<DG.DataFrame> {
     return await grok.data.query('UsageAnalysis:UserInfoByEmailPanel', { email });
   }
@@ -424,10 +432,18 @@ export namespace queries {
     return await grok.data.query('UsageAnalysis:BenchmarkAnalysis', {});
   }
 
+  /**
+   * @param {string} instanceFilter
+   *   choices: ['', 'dev', 'release', 'public']
+   */
   export async function benchmarksDashboard(instanceFilter: string , lastBuildsNum: number , showNotRun?: boolean , showBenchmarks?: boolean , showNotCiCd?: boolean ): Promise<DG.DataFrame> {
     return await grok.data.query('UsageAnalysis:BenchmarksDashboard', { instanceFilter, lastBuildsNum, showNotRun, showBenchmarks, showNotCiCd });
   }
 
+  /**
+   * @param {string} instanceFilter
+   *   choices: ['', 'dev', 'release', 'public', 'release-ec2']
+   */
   export async function testsDashboard(instanceFilter: string , lastBuildsNum: number , versionFilter: string | null, packageFilter: string | null, showNotRun?: boolean , showBenchmarks?: boolean , showNotCiCd?: boolean ): Promise<DG.DataFrame> {
     return await grok.data.query('UsageAnalysis:TestsDashboard', { instanceFilter, lastBuildsNum, versionFilter, packageFilter, showNotRun, showBenchmarks, showNotCiCd });
   }
@@ -436,16 +452,20 @@ export namespace queries {
     return await grok.data.query('UsageAnalysis:ManualTests', { lastBatchesNum });
   }
 
-  export async function stressTestsDashboard(lastBuildsNum: number ): Promise<DG.DataFrame> {
-    return await grok.data.query('UsageAnalysis:StressTestsDashboard', { lastBuildsNum });
+  /**
+   * @param {string} build
+   *   choices: Query("select b.name from builds b where exists (select 1 from stress_tests s where s.build_name = b.name) order by b.build_date desc")
+   */
+  export async function stressTestsRaw(build: string | null): Promise<DG.DataFrame> {
+    return await grok.data.query('UsageAnalysis:StressTestsRaw', { build });
+  }
+
+  export async function stressTestsSummary(lastBuildsNum: number ): Promise<DG.DataFrame> {
+    return await grok.data.query('UsageAnalysis:StressTestsSummary', { lastBuildsNum });
   }
 
   export async function benchmarks(): Promise<DG.DataFrame> {
     return await grok.data.query('UsageAnalysis:Benchmarks', {});
-  }
-
-  export async function stressTests(batch_name: string | null): Promise<DG.DataFrame> {
-    return await grok.data.query('UsageAnalysis:StressTests', { batch_name });
   }
 
   export async function lastBuildsBenchmarksCompare(): Promise<DG.DataFrame> {
@@ -603,8 +623,8 @@ export namespace funcs {
   }
 
   /**
-  Creates JIRA ticket using current error log
-  */
+   * Creates JIRA ticket using current error log
+   */
   export async function createJiraTicket(): Promise<void> {
     return await grok.functions.call('UsageAnalysis:CreateJiraTicket', {});
   }